Rename output to serial-json format.
