+++ main/src/main/resources/application.conf
@@ -18,7 +18,7 @@ outDir = ${rootDir}/output
 # "arizona" (column-based, one file per paper)
 # "cmu" (column-based, one file per paper)
 # "fries" (multiple JSON files per paper)
-# "mentions-json" (one JSON file for all extractions in a paper)
+# "serial-json" (JSON serialization of mentions data structures. LARGE output!)
 # "text" (non-JSON textual format)
 outputTypes = ["fries"]
